OAuth Setup Rollout: Enhancing Website Security

+++ backend/src/server.js
@@ -17,12 +17,29 @@ const getAccessAndBearerTokenUrl = (assess_token) => {
     return `https://www.googleapis.com/oauth2/v1/userinfo?alt=json&access_token=${assess_token}`;
 }
 
-// const updateOrCreateUserFromOauth = async (user) => {
-//     const database = client.db('easyshopper');
-//     const users = database.collection('users');
-//     const existingUser = await
-//         users.findOne({ email: user.email });
+const updateOrCreateUserFromOauth = async (oauthClient) => {
+    const { name,
+            email,
+    } = oauthClient;
 
+    console.log(name, email);
+
+    const database = client.db('easyshopper');
+    const users = database.collection('users');
+
+    const existingUser = await users.findOne({ email });
+
+    if (existingUser) {
+        const result = await users.findOneAndUpdate(
+            { email }, { $set: { name, email} },
+            { returnDocument: 'after' }
+        );
+        return result;
+    } else {
+        const result = await users.insertOne({ name, email });
+        return {email, name, _id: result.insertedId};
+    }
+}
 
 app.get('/api/google/auth/callback', async (req, res) => {
     const { code } = req.query;
@@ -49,7 +66,7 @@ app.get('/api/google/auth/callback', async (req, res) => {
     //TODO: Redirect to the login page with the JWT token - update the url
     fetch(url, requestOptions)
         .then((response) => response.json())
-        //.then((result) => updateOrCreateUserFromOauth(result))
+        .then((result) =>  updateOrCreateUserFromOauth(result))
         .then((user) => {
             console.log(user)
             jwt.sign( {"name":user.name, "email":user.email}, JWTSecret, {expiresIn: '2d'}, (err, token) => {
@@ -68,6 +85,26 @@ app.get('/api/google/auth/callback', async (req, res) => {
 
 app.get('/api/google/url', (req, res) => {
     res.status(200).json({ url: googleOauthURL });
+});
+
+app.get('/api/verifyToken', (req, res) => {
+    const { authorization } = req.headers;
+    if (!authorization) {
+        return res.status(400).json({ error: 'Authorization header is required' });
+    }
+    const token = authorization.split(' ')[1];
+    jwt.verify(token, JWTSecret, (err, decoded) => {
+        if (err) {
+            return res.status(401).json({ error: 'Invalid token' });
+        }
+        const database = client.db('easyshopper');
+        const users = database.collection('users');
+        const user = users.findOne({ email: decoded.email });
+        if (!user) {
+            return res.status(401).json({ error: 'Invalid token' });
+        }
+        return res.status(200).json({ email: decoded.email });
+    });
 });
 
 app.get('/api/products', async (req, res) => {
@@ -89,10 +126,24 @@ app.get('/api/products/:productId', async (req, res) => {
 
 // Route to handle requests for fetching the cart
 app.get('/api/cart', async (req, res) => {
-    const database = client.db('easyshopper');
-    const cart = database.collection('userCarts');
-    const cartItems = await cart.find({}).toArray();
-    res.json({ cart: cartItems });
+    const { authorization } = req.headers;
+    if (!authorization) {
+        return res.status(400).json({ error: 'Authorization header is required' });
+    }
+    try {
+        const token = authorization.split(' ')[1];
+        jwt.verify(token, JWTSecret, async (err, decoded) => {
+            if (err) {
+                return res.status(401).json({ error: 'Invalid token' });
+            }
+            const database = client.db('easyshopper');
+            const cart = database.collection('userCarts');
+            const cartItems = await cart.find({ email: decoded.email }).toArray();
+            return res.json({ cart: cartItems });
+        });
+    } catch (error) {
+        return res.status(500).json({ error: 'Error validating user' });
+    }
 });
 
 // Route to handle requests for adding a product to the cart
@@ -101,27 +152,45 @@ app.post('/api/cart/:productId', async (req, res) => {
     const productId = req.params.productId;
     const quantity = req.body.quantity;
 
-    const database = client.db('easyshopper');
-    const cart = database.collection('userCarts');
-    const product = await cart.findOne({ productId });
-
-    const products = database.collection('products');
-    const productData = await products.findOne({ productId });
-    //TODO: Update the quantity of the product in the products collection, so update by reducing the quantity by the quantity added to the cart.
-    if (productData) {
-        await products.updateOne({ productId }, { $inc: { "inventory.quantity": -quantity } });
-    } else {
-        console.log('Product not found in the products collection');
-    } 
-
-    //TODO: If the quantity is 0, make the product unavailable in the shop, product and the cart page. - Frontend part  
-    if (product) {
-        await cart.updateOne({ productId }, { $inc: { quantity } });
-    } else {
-        await cart.insertOne({ productId, quantity });
+    const { authorization } = req.headers;
+    if (!authorization) {
+        return res.status(400).json({ error: 'Authorization header is required' });
     }
+    try {
+        const token = authorization.split(' ')[1];
+        jwt.verify(token, JWTSecret, async (err, decoded) => {
+            if (err) {
+                return res.status(401).json({ error: 'Invalid token' });
+            }
 
-    res.json({ status: 'ok' });
+            const database = client.db('easyshopper');
+            const cart = database.collection('userCarts');
+            // Add email condition
+            const product = await cart.findOne({ productId, email: decoded.email });
+
+            // Updates the quantity of the product in the products collection
+            const products = database.collection('products');
+            const productData = await products.findOne({ productId });
+            //TODO: Update the quantity of the product in the products collection, so update by reducing the quantity by the quantity added to the cart.
+            if (productData) {
+                await products.updateOne({ productId }, { $inc: { "inventory.quantity": -quantity } });
+            } else {
+                console.log('Product not found in the products collection');
+            }
+
+            //TODO: If the quantity is 0, make the product unavailable in the shop, product and the cart page. - Frontend part
+            if (product) {
+                await cart.updateOne({ productId }, { $inc: { quantity } });
+            } else {
+                await cart.insertOne({ productId, quantity, email: decoded.email });
+            }
+
+            return res.json({ status: "Product added to cart" });
+        });
+
+    } catch (error) {
+        return res.status(500).json({ error: 'Error validating user' });
+    }
 });
 
 // Route to handle requests for removing a product from the cart
@@ -130,26 +199,42 @@ app.post('/api/cart/:productId', async (req, res) => {
 app.delete('/api/cart/:productId', async (req, res) => {
     const productId = req.params.productId;
 
-    const database = client.db('easyshopper');
-    const cart = database.collection('userCarts');
-    const product = await cart.findOne({ productId });
-
-    const products = database.collection('products');
-    const productData = await products.findOne({ productId });
-
-    if (productData) {
-        await products.updateOne({ productId }, { $inc: { "inventory.quantity": + cart.quantity } });
-    } else {
-        console.log('Product not found in the products collection');
+    const { authorization } = req.headers;
+    if (!authorization) {
+        return res.status(400).json({ error: 'Authorization header is required' });
     }
+    try {
+        const token = authorization.split(' ')[1];
+        jwt.verify(token, JWTSecret, async (err, decoded) => {
+            if (err) {
+                return res.status(401).json({ error: 'Invalid token' });
+            }
 
-    if (product) {
-        await cart.deleteOne({ productId });
-    } else {
-        console.log('Product not found in the cart collection');
+            const database = client.db('easyshopper');
+            const cart = database.collection('userCarts');
+            const product = await cart.findOne({ productId, email: decoded.email });
+
+            const products = database.collection('products');
+            const productData = await products.findOne({ productId });
+
+            if (productData) {
+                await products.updateOne({ productId }, { $inc: { "inventory.quantity": + product.quantity } });
+            } else {
+                console.log('Product not found in the products collection');
+            }
+
+            if (product) {
+                await cart.deleteOne({ productId });
+            } else {
+                console.log('Product not found in the cart collection');
+            }
+
+            return res.json({ status: 'Product removed from cart' });
+        });
+
+    } catch (error) {
+        return res.status(500).json({ error: 'Error validating user' });
     }
-
-    res.json({ status: 'ok' });
 });
 
 //TODO: Add GET /api/products/:id/reviews endpoint to get all reviews for a product
